# backend/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Bookstore - Hypermedia applications project \n"
  version: "1.0.0"
  title: "Bookstore"
  contact:
    email: "davide@avanzi.dev"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/v2"
...
schemes:
- "https"
- "http"
paths:
...
  /event:
    get:
      tags:
      - "event"
      summary: "All events, optionally filtered"
      description: "List of events"
      operationId: "getEvents"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20, max is 500."
        required: false
        type: "integer"
      - name: "bookId"
        in: "query"
        description: "Id of the presented book on the event"
        required: false
        type: "integer"
      responses:
        200:
          description: "List of events"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Event"
        404:
          description: "Not found"
      x-swagger-router-controller: "Event"
    post:
      tags:
      - "event"
      summary: "Add a new event"
      description: "Insert a new event in the system - this can only be done by logged\
        \ in ADMIN."
      operationId: "addEvent"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The review object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        200:
          description: "Event added"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Event"
  /event/{eventId}:
    get:
      tags:
      - "event"
      summary: "Find event by ID"
      description: "Returns a event"
      operationId: "getEventById"
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        description: "ID of the event to retrieve"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Event"
        400:
          description: "Invalid input"
        404:
          description: "Not found"
      x-swagger-router-controller: "Event"
    put:
      tags:
      - "event"
      summary: "Update a specific event"
      description: "Update values of a review - this can only be done by the logged\
        \ in ADMIN."
      operationId: "updateEvent"
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        description: "id of the event that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Updated event object"
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        200:
          description: "Event updated"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "Event"
    delete:
      tags:
      - "event"
      summary: "Delete a specific event"
      description: "Delete a event - this can only be done by the logged in ADMIN."
      operationId: "deleteEvent"
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        description: "The id of the event that needs to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Event deleted"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "Event"
...
  /interview:
    get:
      tags:
      - "interview"
      summary: "All interviews, optionally filtered"
      description: "List of interviews"
      operationId: "getInterviews"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20, max is 500."
        required: false
        type: "integer"
      - name: "bookId"
        in: "query"
        description: "Id of the book to get interviews"
        required: false
        type: "integer"
      responses:
        200:
          description: "List of interviews"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Interview"
        404:
          description: "Not found"
      x-swagger-router-controller: "Interview"
    post:
      tags:
      - "interview"
      summary: "Add a new interview"
      description: "Insert a new interview in the system - this can only be done by\
        \ logged users."
      operationId: "addInterview"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The interview object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Interview"
      responses:
        200:
          description: "Interview added"
        403:
          description: "Unauthorized"
        404:
          description: "Not found"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Interview"
  /interview/{interviewId}:
    get:
      tags:
      - "interview"
      summary: "Find interview by ID"
      description: "Insert get a single reservation in the system - this can only\
        \ be done by logged users or admin."
      operationId: "getInterviewById"
      produces:
      - "application/json"
      parameters:
      - name: "interviewId"
        in: "path"
        description: "ID of the interview to retrieve"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Interview"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "Interview"
    put:
      tags:
      - "interview"
      summary: "Update a specific interview"
      description: "Update values of a interview - this can only be done by the logged\
        \ in ADMIN or the user that made it."
      operationId: "updateInterview"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "interviewId"
        in: "path"
        description: "id of the interview that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Interview updated"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "Interview"
    delete:
      tags:
      - "interview"
      summary: "Delete a specific interview"
      description: "Delete a interview - this can only be done by the logged in ADMIN\
        \ or the user that made it."
      operationId: "deleteInterview"
      produces:
      - "application/json"
      parameters:
      - name: "interviewId"
        in: "path"
        description: "The id of the interview that needs to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Interview deleted"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "Interview"
  /theme:
    get:
      tags:
      - "theme"
      summary: "All themes, optionally filtered"
      description: "List of thmes"
      operationId: "getThemes"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20, max is 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "List of themes"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Theme"
        404:
          description: "Not found"
      x-swagger-router-controller: "Theme"
    post:
      tags:
      - "theme"
      summary: "Add a new theme"
      description: "Insert a new theme in the system - this can only be done by logged\
        \ users."
      operationId: "addTheme"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The name object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Theme"
      responses:
        200:
          description: "Theme added"
        403:
          description: "Unauthorized"
        404:
          description: "Not found"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Theme"
  /theme/{themeId}:
    get:
      tags:
      - "theme"
      summary: "Find theme by ID"
      description: "Insert get a single theme in the system"
      operationId: "getThemeById"
      produces:
      - "application/json"
      parameters:
      - name: "themeId"
        in: "path"
        description: "ID of the theme to retrieve"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Theme"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "Theme"
    put:
      tags:
      - "theme"
      summary: "Update a specific theme"
      description: "Update values of a theme - this can only be done by the logged\
        \ in ADMIN."
      operationId: "updateTheme"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "themeId"
        in: "path"
        description: "id of the theme that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Theme updated"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "Theme"
    delete:
      tags:
      - "theme"
      summary: "Delete a specific theme"
      description: "Delete a theme - this can only be done by the logged in ADMIN."
      operationId: "deleteTheme"
      produces:
      - "application/json"
      parameters:
      - name: "themeId"
        in: "path"
        description: "The id of the theme that needs to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Theme deleted"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "Theme"
  /genre:
    get:
      tags:
      - "genre"
      summary: "All genres, optionally filtered"
      description: "List of genres"
      operationId: "getGenres"
      produces:
      - "application/json"
      parameters:
      - name: "offset"
        in: "query"
        description: "Pagination offset. Default is 0"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Maximum number of items per page. Default is 20, max is 500."
        required: false
        type: "integer"
      responses:
        200:
          description: "List of genres"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Genre"
        404:
          description: "Not found"
      x-swagger-router-controller: "Genre"
    post:
      tags:
      - "genre"
      summary: "Add a new genre"
      description: "Insert a new genre in the system - this can only be done by logged\
        \ ADMIN."
      operationId: "addGenre"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The name object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Genre"
      responses:
        200:
          description: "Genre added"
        403:
          description: "Unauthorized"
        404:
          description: "Not found"
        405:
          description: "Invalid input"
      x-swagger-router-controller: "Genre"
  /genre/{genreId}:
    get:
      tags:
      - "genre"
      summary: "Find genre by ID"
      description: "Insert get a single genre in the system"
      operationId: "getGenreById"
      produces:
      - "application/json"
      parameters:
      - name: "genreId"
        in: "path"
        description: "ID of the genre to retrieve"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Genre"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "Genre"
    put:
      tags:
      - "genre"
      summary: "Update a specific genre"
      description: "Update values of a genre - this can only be done by the logged\
        \ in ADMIN."
      operationId: "updateGenre"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "genreId"
        in: "path"
        description: "id of the genre that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Genre updated"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "Genre"
    delete:
      tags:
      - "genre"
      summary: "Delete a specific genre"
      description: "Delete a genre - this can only be done by the logged in ADMIN."
      operationId: "deleteGenre"
      produces:
      - "application/json"
      parameters:
      - name: "genreId"
        in: "path"
        description: "The id of the theme that needs to be deleted"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "Genre deleted"
        400:
          description: "Invalid input"
        403:
          description: "Forbidden"
        404:
          description: "Not found"
      x-swagger-router-controller: "genre"
definitions:
...
  Event:
    type: "object"
    required:
    - "bookId"
    properties:
      id:
        type: "integer"
        format: "int64"
      bookId:
        type: "integer"
        format: "int64"
      title:
        type: "string"
      location:
        type: "string"
      content:
        type: "string"
    xml:
      name: "Event"
    example:
      location: "location"
      id: 0
      title: "title"
      content: "content"
      bookId: 6
...
  Interview:
    type: "object"
    required:
    - "bookId"
    - "id"
    properties:
      id:
        type: "integer"
        format: "int64"
      bookId:
        type: "integer"
        format: "int64"
      title:
        type: "string"
        description: "title of the interview"
      content:
        type: "string"
        description: "message of the interview"
      interviewer:
        type: "string"
        description: "name of the interviewer"
    xml:
      name: "Interview"
    example:
      id: 0
      title: "title"
      content: "content"
      interviewr: "interviewer"
      bookId: 1
  Theme:
    type: "object"
    required:
    - "id"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
        description: "name of the theme"
    xml:
      name: "Interview"
    example:
      id: 0
      name: "Theme"
  Genre:
    type: "object"
    required:
    - "id"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
        description: "name of the genre"
    xml:
      name: "Genre"
    example:
      id: 0
      name: "name"